Fixed issue #4
The login() method of the Salesforce SOAP API failed. This was due to the fact that the security token now has to be appended to the password.

Therefore I added the security token as value of the App.config and changed the parameters of the application. You now can use -t to set the Security Token and -h to set the host. This is a breaking change compared to the original version of this application. Up until now -h was used to display the help.

Additional changes:
* Fixed SSL issue in SalesForceWebDownloader (failed to establish connection)
* Updated regular expressions used to parse the result page (old ones were broken)
* Removed unsused imports
* Unified cases of names of application settings (AppSettingKeys)